\name{AlphabetSoup-package}
\alias{AlphabetSoup-package}
\alias{AlphabetSoup}
\docType{package}
\title{
\packageTitle{AlphabetSoup}
}
\description{
\packageDescription{AlphabetSoup}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlphabetSoup}
\packageIndices{AlphabetSoup}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AlphabetSoup}

Maintainer:  \packageMaintainer{AlphabetSoup}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{s
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
